+++ packages/nimble-components/src/table/components/row/index.ts
@@ -24,9 +24,9 @@ export class TableRow<
     @observable
     public columns: TableColumn[] = [];
 
-    public getCellValue(column: TableColumn): TableCellState<TableRecord> {
+    public getCellState(column: TableColumn): TableCellState<TableRecord> {
         const fieldNames = column.getRecordFieldNames();
-        if (this.hasValidFieldNames(fieldNames)) {
+        if (this.hasValidFieldNames(fieldNames) && this.data) {
             const cellDataValues = fieldNames.map(field => this.data![field]);
             const cellData = Object.fromEntries(
                 column.cellStateDataFieldNames.map((k, i) => [